Create AllTrueBaseCmds.gs
I decided to make the base commands (ls, cat, nmap, etc) here but I figured out a function that will automatically treat user-inputs with the same name as a bin binary as the bin's script and run it. Works regardless of whether it's local or remote. Since i figured out that short cut, I started putting all other complicated commands in this src file. Some of these commands are still a work in progress.